+++ client/package.json
@@ -20,9 +20,12 @@
     "react-router-dom": "^6.10.0"
   },
   "devDependencies": {
-    "@types/react": "^18.0.28",
+    "@mui/types": "^7.2.4",
+    "@types/react": "^18.0.35",
     "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react": "^3.1.0",
+    "@types/node": "^18.15.11",
+    "typescript": "^5.0.4",
     "vite": "^4.2.0"
   }
 }
