+++ client/src/App.tsx
@@ -1,4 +1,5 @@
 import './app.css';
+import React from "react";
 
 function App() {
     return (
