+++ client/src/components/Footer/Footer.tsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 const Footer = () => {
     return (
         <div className="footer">
